Entity Framework kullanarak veritabanı bağlantısı yapıldı

+++ Entities/Concrete/Car.cs
@@ -13,7 +13,7 @@ namespace Entities.Concrete
         public int BrandId { get; set; }
         public int ModelYear { get; set; }
         public string Description { get; set; }
-        public short ColorId { get; set; }
+        public int ColorId { get; set; }
         public decimal DailyPrice { get; set; }
     }
 }
